First revision of implement GPS
- renamed SoftwareSerial "mySerial" to "gpsSerialPort"
- removed serial USB initialisation as already initialised
- moved initialisation stuff to gpio_init()
- moved GPS reading setup to line 37
- removed if statement of GPSECHO
- changed daq_GPSVehicleSpeed_t from int32_t to float32_t

+++ cmn/daq_common.h
@@ -11,7 +11,7 @@ typedef int32_t daq_Strain_t;
 typedef uint32_t daq_BatteryV_t;
 typedef int32_t daq_ThrottlePos_t;
 typedef int32_t daq_FuelPressure_t;
-typedef int32_t daq_GPSVehicleSpeed_t;
+typedef float32_t daq_GPSVehicleSpeed_t;
 
 typedef struct WheelSpeed_s
 {
